// src/lib/audit/schemas/objectifs.ts
export interface ObjectifsData {
  objectifs6Mois: string;
  objectifs12Mois: string;
  objectifCA: string;
  objectifRentabilite: string;
  objectifOrganisation: string;
  objectifRecrutement: string;
  objectifFormation: string;
  objectifInvestissement: string;
  resultatPrioritaire: string;
  mesureSucces: string;
}

export const OBJECTIFS_EMPTY: ObjectifsData = {
  objectifs6Mois: "",
  objectifs12Mois: "",
  objectifCA: "",
  objectifRentabilite: "",
  objectifOrganisation: "",
  objectifRecrutement: "",
  objectifFormation: "",
  objectifInvestissement: "",
  resultatPrioritaire: "",
  mesureSucces: "",
};